requirements.txt
python-multipart
pillow
ultralytics==8.0.197
python-dotenv
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
matplotlib==3.10.0
numpy==1.23.5
